+++ Cql/CodeGeneration.NET/CodeGeneration.NET.csproj
@@ -29,7 +29,7 @@
 		<InternalsVisibleTo Include="Hl7.Cql.Packager" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="CoreTests" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="CqlToElmTests" Key="$(LibraryPKHash)" />
-		<InternalsVisibleTo Include="CqlSdkPrototype" Key="$(LibraryPKHash)" />
+		<InternalsVisibleTo Include="Hl7.Cql.Invocation" Key="$(LibraryPKHash)" />
 	</ItemGroup>
 
 </Project>
