+++ Cql/CoreTests/Tuples/CqlTupleTests.cs
@@ -1,8 +1,8 @@
 ﻿#nullable enable
-using CqlSdkPrototype.Invocation;
-using CqlSdkPrototype.Invocation.Extensions;
 using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Fhir;
+using Hl7.Cql.Invocation.Toolkit;
+using Hl7.Cql.Invocation.Toolkit.Extensions;
 using Hl7.Cql.Primitives;
 using Hl7.Cql.Runtime;
 using Hl7.Cql.Runtime.Serialization;
