+++ Cql/Cql.Abstractions/Cql.Abstractions.csproj
@@ -33,6 +33,7 @@
 		<InternalsVisibleTo Include="CoreTests" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="Hl7.Cql.CodeGeneration.NET" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="CqlSdkPrototype" Key="$(LibraryPKHash)" />
+		<InternalsVisibleTo Include="Hl7.Cql.Invocation" Key="$(LibraryPKHash)" />
 	</ItemGroup>
 
 	<ItemGroup>
