+++ Cql/Cql.Invocation/Cql.Invocation.csproj
@@ -3,6 +3,7 @@
 	<Import Project="..\..\cql-sdk.props" />
 
 	<PropertyGroup>
+		<ImplicitUsings>enable</ImplicitUsings>
 		<AssemblyName>Hl7.Cql.Invocation</AssemblyName>
 		<RootNamespace>Hl7.Cql.Invocation</RootNamespace>
 		<Product>CQL Library Invocation</Product>
@@ -19,8 +20,15 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\CodeGeneration.NET\CodeGeneration.NET.csproj" />
+		<ProjectReference Include="..\Cql.Compiler\Cql.Compiler.csproj" />
+		<ProjectReference Include="..\Cql.CqlToElm\Cql.CqlToElm.csproj" />
+		<ProjectReference Include="..\Cql.Firely\Cql.Firely.csproj" />
+		<ProjectReference Include="..\Cql.Primitives\Cql.Primitives.csproj" />
+		<ProjectReference Include="..\Cql.Packaging\Cql.Packaging.csproj" />
 	</ItemGroup>
 
 </Project>
