+++ Cql/Cql.Invocation/GlobalUsings.cs
@@ -8,7 +8,9 @@
 
 // Global using directives
 
+global using System;
 global using System.Collections.Concurrent;
+global using System.Collections.Generic;
 global using System.Collections.Immutable;
 global using System.Diagnostics;
 global using System.Diagnostics.CodeAnalysis;
