+++ Cql/Cql.Invocation/PublicAPI.Unshipped.txt
@@ -1,0 +1,39 @@
+﻿#nullable enable
+abstract Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.Invoke(Hl7.Cql.Runtime.CqlContext! cqlContext) -> object?
+abstract Hl7.Cql.Invocation.Toolkit.LibraryInvoker.Definitions.get -> System.Collections.Generic.IReadOnlyDictionary<string!, Hl7.Cql.Invocation.Toolkit.DefinitionInvoker!>!
+abstract Hl7.Cql.Invocation.Toolkit.LibraryInvoker.LibraryIdentifier.get -> Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.DefinitionInvoker(string! definitionName, Hl7.Cql.Abstractions.ILibrary! library, System.Reflection.MethodInfo! methodInfo, System.Collections.Generic.IReadOnlyDictionary<string!, string!>! tagValuesByName, string? valueSetId) -> void
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.DefinitionName.get -> string!
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.InvokeMethod(params object?[]! parameters) -> object?
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.ReturnType.get -> System.Type!
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.TagValuesByName.get -> System.Collections.Generic.IReadOnlyDictionary<string!, string!>!
+Hl7.Cql.Invocation.Toolkit.DefinitionInvoker.ValueSetId.get -> string?
+Hl7.Cql.Invocation.Toolkit.Extensions.CqlToolkitExtensions
+Hl7.Cql.Invocation.Toolkit.Extensions.ElmToolkitExtensions
+Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions
+Hl7.Cql.Invocation.Toolkit.Extensions.LibraryInvokerExtensions
+Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions
+Hl7.Cql.Invocation.Toolkit.InvocationToolkit
+Hl7.Cql.Invocation.Toolkit.InvocationToolkit.AddAssemblyBinaries(System.Collections.Generic.IEnumerable<Hl7.Cql.CodeGeneration.NET.AssemblyBinary!>! assemblyBinary) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
+Hl7.Cql.Invocation.Toolkit.InvocationToolkit.AssemblyBinaries.get -> System.Collections.Generic.IReadOnlySet<Hl7.Cql.CodeGeneration.NET.AssemblyBinary!>!
+Hl7.Cql.Invocation.Toolkit.InvocationToolkit.CreateLibrarySetInvoker(string! name = "") -> Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker!
+Hl7.Cql.Invocation.Toolkit.InvocationToolkit.InvocationToolkit(Microsoft.Extensions.Logging.ILoggerFactory? loggerFactory = null) -> void
+Hl7.Cql.Invocation.Toolkit.LibraryInvoker
+Hl7.Cql.Invocation.Toolkit.LibraryInvoker.LibraryInvoker() -> void
+Hl7.Cql.Invocation.Toolkit.LibraryInvokerOnInstance
+Hl7.Cql.Invocation.Toolkit.LibraryInvokerOnInstance.Library.get -> Hl7.Cql.Abstractions.ILibrary!
+Hl7.Cql.Invocation.Toolkit.LibraryInvokerOnInstance.LibraryInvokerOnInstance(Hl7.Cql.Abstractions.ILibrary! library) -> void
+Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker
+Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker.Dispose() -> void
+Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker.LibraryInvokers.get -> System.Collections.Generic.IReadOnlyDictionary<Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier, Hl7.Cql.Invocation.Toolkit.LibraryInvoker!>!
+override Hl7.Cql.Invocation.Toolkit.LibraryInvokerOnInstance.LibraryIdentifier.get -> Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier
+static Hl7.Cql.Invocation.Toolkit.Extensions.CqlToolkitExtensions.CreateLibrarySetInvoker(this Hl7.Cql.CqlToElm.Toolkit.CqlToolkit! cqlToolkit, Hl7.Cql.Abstractions.Mutator<Hl7.Cql.CodeGeneration.NET.Toolkit.ElmToolkitConfig!>? reconfigureElmToolkitConfig = null, string! name = "") -> Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker!
+static Hl7.Cql.Invocation.Toolkit.Extensions.ElmToolkitExtensions.CreateInvocationToolkit(this Hl7.Cql.CodeGeneration.NET.Toolkit.ElmToolkit! elmToolkit) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
+static Hl7.Cql.Invocation.Toolkit.Extensions.ElmToolkitExtensions.CreateLibrarySetInvoker(this Hl7.Cql.CodeGeneration.NET.Toolkit.ElmToolkit! elmToolkit, string! name = "") -> Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker!
+static Hl7.Cql.Invocation.Toolkit.Extensions.InvocationToolkitExtensions.AddAssemblyBinaries(this Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, params Hl7.Cql.CodeGeneration.NET.AssemblyBinary![]! assemblyBinary) -> Hl7.Cql.Invocation.Toolkit.InvocationToolkit!
+static Hl7.Cql.Invocation.Toolkit.Extensions.LibraryInvokerExtensions.EnumerateLibraryDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibraryInvoker! libraryInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, System.Func<object?>! getResult)>!
+static Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions.EnumerateLibraryDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker! librarySetInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext, Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier libraryIdentifier) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.LibraryInvoker! libraryInvoker, Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, System.Func<object?>! getResult)>!
+static Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions.EnumerateLibrarySetDefinitionsResults(this Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker! librarySetInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext) -> System.Collections.Generic.IEnumerable<(Hl7.Cql.Invocation.Toolkit.LibraryInvoker! libraryInvoker, Hl7.Cql.Invocation.Toolkit.DefinitionInvoker! definitionInvoker, System.Func<object?>! getResult)>!
+static Hl7.Cql.Invocation.Toolkit.Extensions.LibrarySetInvokerExtensions.GetLibraryDefinitionResult(this Hl7.Cql.Invocation.Toolkit.LibrarySetInvoker! librarySetInvoker, Hl7.Cql.Runtime.CqlContext! cqlContext, Hl7.Cql.Runtime.CqlVersionedLibraryIdentifier libraryIdentifier, string! definitionName) -> object?
+static Hl7.Cql.Invocation.Toolkit.LibraryInvoker.TryCreateFromType(Hl7.Cql.Invocation.Toolkit.InvocationToolkit! invocationToolkit, System.Type! libraryType, out Hl7.Cql.Invocation.Toolkit.LibraryInvoker? libraryInvoker) -> bool
