+++ Cql/Cql.Invocation/Toolkit/Extensions/CqlToolkitExtensions.cs
@@ -11,10 +11,20 @@ using Hl7.Cql.CodeGeneration.NET.Toolkit;
 using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.CqlToElm.Toolkit;
 
-namespace CqlSdkPrototype.Invocation.Extensions;
+namespace Hl7.Cql.Invocation.Toolkit.Extensions;
 
+/// <summary>
+/// Provides extension methods for the <see cref="CqlToolkit"/> class.
+/// </summary>
 public static class CqlToolkitExtensions
 {
+    /// <summary>
+    /// Creates a <see cref="LibrarySetInvoker"/> from the specified <see cref="CqlToolkit"/>.
+    /// </summary>
+    /// <param name="cqlToolkit">The CQL toolkit to use for creating the library set invoker.</param>
+    /// <param name="reconfigureElmToolkitConfig">An optional mutator to reconfigure the ELM toolkit configuration.</param>
+    /// <param name="name">The name of the library set invoker.</param>
+    /// <returns>A new instance of <see cref="LibrarySetInvoker"/>.</returns>
     public static LibrarySetInvoker CreateLibrarySetInvoker(
         this CqlToolkit cqlToolkit,
         Mutator<ElmToolkitConfig>? reconfigureElmToolkitConfig = null,
