+++ Cql/Cql.Packaging/Cql.Packaging.csproj
@@ -30,10 +30,4 @@
 		<InternalsVisibleTo Include="CqlToElmTests" Key="$(LibraryPKHash)" />
 		<InternalsVisibleTo Include="CqlSdkPrototype" Key="$(LibraryPKHash)" />
 	</ItemGroup>
-
-	<ItemGroup>
-		<Compile Update="CqlPackagingPipelineOptions..*.cs">
-			<DependentUpon>CqlPackagingPipelineOptions..cs</DependentUpon>
-		</Compile>
-	</ItemGroup>
 </Project>
