+++ Cql/CqlToElmTests/(tests)/FunctionDefinitionTest.cs
@@ -1,6 +1,6 @@
-using CqlSdkPrototype.Invocation.Extensions;
 using Hl7.Cql.CodeGeneration.NET.Toolkit.Extensions;
 using Hl7.Cql.Elm;
+using Hl7.Cql.Invocation.Toolkit.Extensions;
 using Hl7.Cql.Runtime;
 
 namespace Hl7.Cql.CqlToElm.Test
