+++ Demo/CLI/CLI.csproj
@@ -10,6 +10,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\..\Cql\Cql.Firely\Cql.Firely.csproj" />
+		<ProjectReference Include="..\..\Cql\Cql.Invocation\Cql.Invocation.csproj" />
 		<ProjectReference Include="..\..\Cql\Cql.Packaging\Cql.Packaging.csproj" />
 		<ProjectReference Include="..\..\Cql\CqlSdkPrototype\CqlSdkPrototype.csproj" />
 	</ItemGroup>
