+++ Demo/CLI/Helpers/LibraryExtensions.cs
@@ -6,10 +6,10 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using CqlSdkPrototype.Invocation;
 using Hl7.Cql.Abstractions;
 using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Fhir;
+using Hl7.Cql.Invocation.Toolkit;
 using Hl7.Cql.ValueSets;
 using Hl7.Fhir.Model;
 using Hl7.Cql.Packaging;
