+++ Demo/CLI/Helpers/ResourceHelper.cs
@@ -6,10 +6,10 @@
  * available at https://raw.githubusercontent.com/FirelyTeam/firely-cql-sdk/main/LICENSE
  */
 
-using CqlSdkPrototype.Invocation;
 using Dumpify;
 using Hl7.Cql.Conversion;
 using Hl7.Cql.Fhir;
+using Hl7.Cql.Invocation.Toolkit;
 using Hl7.Cql.Packaging;
 using Hl7.Cql.Primitives;
 using Hl7.Cql.ValueSets;
