+++ Demo/CqlApiExamples/CqlApiExamples.csproj
@@ -8,6 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Cql\Cql.Invocation\Cql.Invocation.csproj" />
     <ProjectReference Include="..\..\Cql\CqlSdkPrototype\CqlSdkPrototype.csproj" />
   </ItemGroup>
