+++ Demo/Test.Measures.Demo/MeasuresTest.cs
@@ -6,9 +6,9 @@ using Hl7.Cql.Compiler;
 using CLI.Helpers;
 using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Runtime;
-using CqlSdkPrototype.Invocation;
-using CqlSdkPrototype.Invocation.Extensions;
 using Hl7.Cql.CodeGeneration.NET.Toolkit;
+using Hl7.Cql.Invocation.Toolkit;
+using Hl7.Cql.Invocation.Toolkit.Extensions;
 
 namespace Test
 {
